.requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
beautifulsoup4
python-dotenv
